nightscout unfiltered/filtered values are integer values

+++ alrightypump-cloud/src/main/kotlin/com/kludgenics/alrightypump/cloud/nightscout/records/Sgv.kt
@@ -4,8 +4,8 @@ interface Sgv : NightscoutEntry {
     override val type: String get() = "sgv"
     val sgv: Int
     val direction: String
-    val filtered: Double?
-    val unfiltered: Double?
+    val filtered: Int?
+    val unfiltered: Int?
     val rssi: Int?
     val noise: Int?
 }
